# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Restaurant Reservation API",
        "description": "API para gestionar reservas de restaurante",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/reservations": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Get all reservations",
                "operationId": "da6589397b0b417ba1d37aec5dea0f04",
                "responses": {
                    "200": {
                        "description": "List of all reservations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Reservation"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Create a new reservation",
                "operationId": "05a1e800577b4205cefc0a1f68d16bd2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "fecha": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-06-15"
                                    },
                                    "hora": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "19:30:00"
                                    },
                                    "numero_de_personas": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "comensal_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "mesa_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reservation created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reservation"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/reservations/{id}": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Get a reservation by ID",
                "operationId": "019b472459b1dce69e54de66668985a9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the reservation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reservation details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reservation"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reservation not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Update a reservation",
                "operationId": "6ae574cf78fbae4335dc39c0b00fe78b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the reservation to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "fecha": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-06-16"
                                    },
                                    "hora": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "20:00:00"
                                    },
                                    "numero_de_personas": {
                                        "type": "integer",
                                        "example": 6
                                    },
                                    "comensal_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "mesa_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reservation updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reservation"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reservation not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Delete a reservation",
                "operationId": "f4d816f69444320e38d4168b5e69a4c8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the reservation to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reservation deleted successfully"
                    },
                    "404": {
                        "description": "Reservation not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Reservation": {
                "title": "Reservation Model",
                "description": "Reservation model representing a table reservation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "fecha": {
                        "type": "string",
                        "format": "date",
                        "example": "2023-06-15"
                    },
                    "hora": {
                        "type": "string",
                        "format": "time",
                        "example": "19:30:00"
                    },
                    "numero_de_personas": {
                        "type": "integer",
                        "example": 4
                    },
                    "comensal_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "comensal": {
                        "type": "object",
                        "nullable": true
                    },
                    "mesa_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "mesa": {
                        "type": "object",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-06-15T19:30:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-06-15T19:30:00.000000Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}